+++ R/zzz.R
@@ -17,6 +17,7 @@
     roboplot.font.title = list(size = 17, family = "Arial, sans-serif", path = NULL, color = "#696969", bold = T, .font = "Arial", .fallback = "Arial"),
     roboplot.height = 550,
     roboplot.infobox = list(background = "#E8E8E8", border = "black", font = "#696969", border_width = 1),
+    roboplot.labels = list(bar = "none", pie = "percent", scatter = "none"),
     roboplot.legend.tidy = FALSE,
     roboplot.linewidth = 2,
     roboplot.locale = list(