+++ NEWS.md
@@ -1,3 +1,8 @@
+# roboplotr 2.5.0
+
+* Adds `label` param to `roboplot()` along with the correspondint `set_label()` function for trace labeling control.
+* Fixes `roboplot(plot_mode = "horizontalfill")` not sorting the bars correclty by `color`.
+
 # roboplotr 2.4.7
 
 * Adds `zoom` parameter to `set_modebar()` to control the level of zoom of images downloaded from `roboplot()` modebars.